requirements/dev.txt
-r base.txt

django-dotenv
django-debug-toolbar
